[atan2] fixed undefined behavior

+++ src/binary64/atan2/tint.h
@@ -380,7 +380,7 @@ static inline void tint_fromd (tint_t *a, double x)
 // copied from ../exp/exp.c
 static inline double as_ldexp(double x, int64_t i){
 #ifdef __x86_64__
-    __m128i sb; sb[0] = i<<52;
+  __m128i sb; sb[0] = (uint64_t)i<<52;
 #if defined(__clang__)
     __m128d r = _mm_set_sd(x);
 #else
